Change android fix for fts backward compatibility patch of sqlite3.c for newer sqlite version
From earlier version of Android, patch of FTS3 backwards compatibility is kept until now.
But, this patch doesn't work well on newer version of sqlite3 (3.7.13~) when it was used.
Currently, sqlite3Fts3Init function is called three times to register "fts1", "fts2" and "fts3" with the same module.
It worked very well on earlier version of sqlite but not for newer version.
hashDestroy() function makes an error when it is called after the first destroy of fts3Module.
sqlite3Fts3Init() function initializes "fts4" module and it is called three times to initialize "fts1" and "fts2".

This patch contains changed source code of sqlite3Fts3Init() function to be called once.

[written by Yongil Jang <yi.jang@lge.com>]

Change-Id: I55f522b55f421f73e83508cfc5e08514bd7da925